modified scoring system
The achievements can generate positive or negative numbers, it got two
messages in the "inbox" and you get -2 points for a post, the result
would be zero and so you would not be notified. Anything, as negatives,
positives and medals will always be "1".

From now on if you have 2 messages in inbox and -2 points of a post,
will be "2+1", which will be equal to 3 notifications.

However when you open the popup the system continues as it always has,
achievements tab still show positive and negative points normally.